move runner creation to adhoc_process_support
