+++ src/python/pants/core/util_rules/adhoc_process_support.py
@@ -14,10 +14,12 @@ from pants.build_graph.address import Address
 from pants.core.goals.package import BuiltPackage, EnvironmentAwarePackageRequest, PackageFieldSet
 from pants.core.goals.run import RunFieldSet, RunInSandboxRequest
 from pants.core.target_types import FileSourceField
+from pants.core.util_rules.environments import EnvironmentNameRequest
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
 from pants.core.util_rules.system_binaries import BashBinary
 from pants.engine.addresses import Addresses, UnparsedAddressInputs
 from pants.engine.env_vars import EnvironmentVars, EnvironmentVarsRequest
+from pants.engine.environment import EnvironmentName
 from pants.engine.fs import (
     EMPTY_DIGEST,
     CreateDigest,
@@ -27,13 +29,14 @@ from pants.engine.fs import (
     MergeDigests,
     Snapshot,
 )
-from pants.engine.internals.native_engine import RemovePrefix
+from pants.engine.internals.native_engine import AddressInput, RemovePrefix
 from pants.engine.process import FallibleProcessResult, Process, ProcessResult, ProductDescription
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import (
     FieldSetsPerTarget,
     FieldSetsPerTargetRequest,
     SourcesField,
+    Target,
     Targets,
     TransitiveTargets,
     TransitiveTargetsRequest,
@@ -89,6 +92,24 @@ class RunnableDependencies:
     immutable_input_digests: Mapping[str, Digest]
     append_only_caches: Mapping[str, str]
     extra_env: Mapping[str, str]
+
+
+@dataclass(frozen=True)
+class ToolRunnerRequest:
+    runnable_address_str: str
+    args: tuple[str, ...]
+    execution_dependencies: tuple[str, ...]
+    runnable_dependencies: tuple[str, ...]
+    target: Target
+
+
+@dataclass(frozen=True)
+class ToolRunner:
+    digest: Digest
+    args: tuple[str, ...]
+    extra_env: Mapping[str, str]
+    append_only_caches: Mapping[str, str]
+    immutable_input_digests: Mapping[str, Digest]
 
 
 #
@@ -358,6 +379,93 @@ def _runnable_dependency_shim(
         exec {binary} "$@"
         """
     ).encode()
+
+
+@rule
+async def create_tool_runner(
+    request: ToolRunnerRequest,
+) -> ToolRunner:
+    runnable_address = await Get(
+        Address,
+        AddressInput,
+        AddressInput.parse(
+            request.runnable_address_str,
+            relative_to=request.target.address.spec_path,
+            description_of_origin=f"Runnable target for {request.target.address.spec_path}",
+        ),
+    )
+
+    addresses = Addresses((runnable_address,))
+    addresses.expect_single()
+
+    runnable_targets = await Get(Targets, Addresses, addresses)
+
+    target = runnable_targets[0]
+
+    run_field_sets, environment_name, execution_environment = await MultiGet(
+        Get(FieldSetsPerTarget, FieldSetsPerTargetRequest(RunFieldSet, runnable_targets)),
+        Get(EnvironmentName, EnvironmentNameRequest, EnvironmentNameRequest.from_target(target)),
+        Get(
+            ResolvedExecutionDependencies,
+            ResolveExecutionDependenciesRequest(
+                address=runnable_address,
+                execution_dependencies=request.execution_dependencies,
+                runnable_dependencies=request.runnable_dependencies,
+            ),
+        ),
+    )
+
+    run_field_set: RunFieldSet = run_field_sets.field_sets[0]
+
+    run_request = await Get(
+        RunInSandboxRequest, {environment_name: EnvironmentName, run_field_set: RunFieldSet}
+    )
+
+    dependencies_digest = execution_environment.digest
+    runnable_dependencies = execution_environment.runnable_dependencies
+
+    extra_env: dict[str, str] = dict(run_request.extra_env or {})
+    extra_path = extra_env.pop("PATH", None)
+
+    extra_sandbox_contents = []
+
+    extra_sandbox_contents.append(
+        ExtraSandboxContents(
+            EMPTY_DIGEST,
+            extra_path,
+            run_request.immutable_input_digests or FrozenDict(),
+            run_request.append_only_caches or FrozenDict(),
+            run_request.extra_env or FrozenDict(),
+        )
+    )
+
+    if runnable_dependencies:
+        extra_sandbox_contents.append(
+            ExtraSandboxContents(
+                EMPTY_DIGEST,
+                f"{{chroot}}/{runnable_dependencies.path_component}",
+                runnable_dependencies.immutable_input_digests,
+                runnable_dependencies.append_only_caches,
+                runnable_dependencies.extra_env,
+            )
+        )
+
+    merged_extras, main_digest = await MultiGet(
+        Get(ExtraSandboxContents, MergeExtraSandboxContents(tuple(extra_sandbox_contents))),
+        Get(Digest, MergeDigests((dependencies_digest, run_request.digest))),
+    )
+
+    extra_env = dict(merged_extras.extra_env)
+    if merged_extras.path:
+        extra_env["PATH"] = merged_extras.path
+
+    return ToolRunner(
+        digest=main_digest,
+        args=run_request.args + tuple(request.args),
+        extra_env=FrozenDict(extra_env),
+        append_only_caches=merged_extras.append_only_caches,
+        immutable_input_digests=merged_extras.immutable_input_digests,
+    )
 
 
 @rule
